# contracts/openapi/ai-service.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Resume Analyzer - AI Service",
    "description": "Internal AI service for CV parsing, extraction, scoring, matching, and Ollama orchestration.",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Health check",
        "description": "Return health status and model availability.",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/extract-features": {
      "post": {
        "tags": [
          "Extraction"
        ],
        "summary": "Extract resume features from parsed document",
        "description": "Extract candidate metadata, canonical skills, and predicted field.",
        "operationId": "extract_resume_features_api_extract_features_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParsedDocument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeFeatures"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analyze-resume": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze resume PDF",
        "description": "Extract features, classify field, score resume, and provide recommendations.",
        "operationId": "analyze_resume_api_analyze_resume_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_resume_api_analyze_resume_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeAnalysisResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analyze-match": {
      "post": {
        "tags": [
          "Matching"
        ],
        "summary": "Match resume with job description",
        "description": "Match resume features against job description skills and ATS criteria.",
        "operationId": "analyze_match_api_analyze_match_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_match_api_analyze_match_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AiMetadata": {
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/AiProvider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "usedFallback": {
            "type": "boolean",
            "title": "Usedfallback"
          },
          "processingMs": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Processingms"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model",
          "usedFallback",
          "processingMs"
        ],
        "title": "AiMetadata"
      },
      "AiProvider": {
        "type": "string",
        "enum": [
          "OLLAMA",
          "RULE_BASED"
        ],
        "title": "AiProvider"
      },
      "Body_analyze_match_api_analyze_match_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Resume PDF file (max 5 MB)"
          },
          "jobDescription": {
            "type": "string",
            "minLength": 50,
            "title": "Jobdescription",
            "description": "Job description text"
          },
          "targetRole": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetrole",
            "description": "Target role name (optional)"
          }
        },
        "type": "object",
        "required": [
          "file",
          "jobDescription"
        ],
        "title": "Body_analyze_match_api_analyze_match_post"
      },
      "Body_analyze_resume_api_analyze_resume_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Resume PDF file (max 5 MB)"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_resume_api_analyze_resume_post"
      },
      "FieldEnum": {
        "type": "string",
        "enum": [
          "Data Science",
          "Web Development",
          "Android Development",
          "iOS Development",
          "UI/UX",
          "Unknown"
        ],
        "title": "FieldEnum"
      },
      "FieldEvidence": {
        "properties": {
          "field": {
            "$ref": "#/components/schemas/FieldEnum"
          },
          "matchedSkills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matchedskills"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "field",
          "confidence"
        ],
        "title": "FieldEvidence"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "healthy"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "qwen3:4b"
          },
          "ollamaReachable": {
            "type": "boolean",
            "title": "Ollamareachable",
            "default": false
          }
        },
        "type": "object",
        "title": "HealthResponse"
      },
      "MatchResult": {
        "properties": {
          "fileName": {
            "type": "string",
            "title": "Filename"
          },
          "targetRole": {
            "type": "string",
            "title": "Targetrole"
          },
          "matchScore": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Matchscore"
          },
          "matchedSkills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matchedskills"
          },
          "missingSkills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missingskills"
          },
          "atsKeywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 15,
            "title": "Atskeywords"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 6,
            "title": "Strengths"
          },
          "weaknesses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 6,
            "title": "Weaknesses"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Recommendations"
          },
          "ai": {
            "$ref": "#/components/schemas/AiMetadata"
          }
        },
        "type": "object",
        "required": [
          "fileName",
          "targetRole",
          "matchScore",
          "ai"
        ],
        "title": "MatchResult"
      },
      "ParsedDocument": {
        "properties": {
          "fileName": {
            "type": "string",
            "title": "Filename"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "pageCount": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Pagecount"
          },
          "sizeBytes": {
            "type": "integer",
            "maximum": 5242880.0,
            "minimum": 1.0,
            "title": "Sizebytes"
          }
        },
        "type": "object",
        "required": [
          "fileName",
          "text",
          "pageCount",
          "sizeBytes"
        ],
        "title": "ParsedDocument"
      },
      "ResumeAnalysisResult": {
        "properties": {
          "fileName": {
            "type": "string",
            "title": "Filename"
          },
          "candidateName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidatename"
          },
          "candidateEmail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidateemail"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "predictedField": {
            "$ref": "#/components/schemas/FieldEnum"
          },
          "fieldEvidence": {
            "items": {
              "$ref": "#/components/schemas/FieldEvidence"
            },
            "type": "array",
            "title": "Fieldevidence"
          },
          "resumeScore": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Resumescore"
          },
          "scoreBreakdown": {
            "$ref": "#/components/schemas/ScoreBreakdown"
          },
          "recommendedSkills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Recommendedskills"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Recommendations"
          },
          "ai": {
            "$ref": "#/components/schemas/AiMetadata"
          }
        },
        "type": "object",
        "required": [
          "fileName",
          "predictedField",
          "resumeScore",
          "scoreBreakdown",
          "ai"
        ],
        "title": "ResumeAnalysisResult"
      },
      "ResumeFeatures": {
        "properties": {
          "candidateName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidatename"
          },
          "candidateEmail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidateemail"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "predictedField": {
            "$ref": "#/components/schemas/FieldEnum",
            "default": "Unknown"
          },
          "fieldEvidence": {
            "items": {
              "$ref": "#/components/schemas/FieldEvidence"
            },
            "type": "array",
            "title": "Fieldevidence"
          }
        },
        "type": "object",
        "title": "ResumeFeatures"
      },
      "ScoreBreakdown": {
        "properties": {
          "contact": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Contact"
          },
          "summary": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Summary"
          },
          "skills": {
            "type": "integer",
            "maximum": 15.0,
            "minimum": 0.0,
            "title": "Skills"
          },
          "education": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Education"
          },
          "experience": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Experience"
          },
          "projects": {
            "type": "integer",
            "maximum": 15.0,
            "minimum": 0.0,
            "title": "Projects"
          },
          "achievementsCertifications": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Achievementscertifications"
          },
          "quantifiedImpact": {
            "type": "integer",
            "maximum": 15.0,
            "minimum": 0.0,
            "title": "Quantifiedimpact"
          },
          "total": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "contact",
          "summary",
          "skills",
          "education",
          "experience",
          "projects",
          "achievementsCertifications",
          "quantifiedImpact",
          "total"
        ],
        "title": "ScoreBreakdown"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}